nass-1329: merge changes
